refactor: move current-user-related services to a module
this eases tests and encapsulates dependencies
