+++ frontend/src/app/home/home.component.ts
@@ -2,7 +2,7 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs/Subscription';
 
 import { UserModel } from '../models/user.model';
-import { UserService } from '../user.service';
+import { CurrentUserService } from '../current-user/current-user.service';
 
 @Component({
   selector: 'gl-home',
@@ -14,11 +14,11 @@ export class HomeComponent implements OnInit, OnDestroy {
   user: UserModel;
   userEventsSubscription: Subscription;
 
-  constructor(private userService: UserService) {
+  constructor(private currentUserService: CurrentUserService) {
   }
 
   ngOnInit() {
-    this.userEventsSubscription = this.userService.userEvents.subscribe(user => this.user = user);
+    this.userEventsSubscription = this.currentUserService.userEvents.subscribe(user => this.user = user);
   }
 
   ngOnDestroy() {
