+++ frontend/src/app/password-change/password-change.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { UserService } from '../user.service';
 import { Router } from '@angular/router';
+import { CurrentUserService } from '../current-user/current-user.service';
 
 @Component({
   selector: 'gl-password-change',
@@ -16,12 +16,12 @@ export class PasswordChangeComponent {
 
   passwordChangeFailed = false;
 
-  constructor(private userService: UserService, private router: Router) { }
+  constructor(private currentUserService: CurrentUserService, private router: Router) { }
 
   changePassword() {
     this.passwordChangeFailed = false;
-    this.userService.checkPassword(this.model.oldPassword)
-      .switchMap(() => this.userService.changePassword(this.model.newPassword))
+    this.currentUserService.checkPassword(this.model.oldPassword)
+      .switchMap(() => this.currentUserService.changePassword(this.model.newPassword))
       .subscribe(
         () => this.router.navigate(['/']),
         () => this.passwordChangeFailed = true);
