+++ frontend/src/app/user-edit/user-edit.component.spec.ts
@@ -3,7 +3,6 @@ import { async, TestBed } from '@angular/core/testing';
 import { UserEditComponent } from './user-edit.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ActivatedRoute, Router } from '@angular/router';
-import { UserService } from '../user.service';
 import { Observable } from 'rxjs/Observable';
 import { UserModel } from '../models/user.model';
 import { NgModule } from '@angular/core';
@@ -11,15 +10,15 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { JwtInterceptorService } from '../jwt-interceptor.service';
 import { ErrorService } from '../error.service';
+import { UserService } from '../user.service';
 
 describe('UserEditComponent', () => {
 
   @NgModule({
     imports: [CommonModule, HttpClientModule, FormsModule, RouterTestingModule, NgbModule.forRoot()],
     declarations: [UserEditComponent],
-    providers: [UserService, JwtInterceptorService, ErrorService]
+    providers: [UserService, ErrorService]
   })
   class TestModule {}
