+++ frontend/src/app/users-resolver.service.spec.ts
@@ -2,14 +2,13 @@ import { TestBed } from '@angular/core/testing';
 
 import { UsersResolverService } from './users-resolver.service';
 import { HttpClientModule } from '@angular/common/http';
-import { UserService } from './user.service';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
-import { JwtInterceptorService } from './jwt-interceptor.service';
+import { UserService } from './user.service';
 
 describe('UsersResolverService', () => {
   beforeEach(() => TestBed.configureTestingModule({
-    providers: [UsersResolverService, UserService, JwtInterceptorService],
+    providers: [UsersResolverService, UserService],
     imports: [HttpClientModule]
   }));
